[finsh][cmd] add 'console' command to config console
