+++ bsp/qemu-virt64-aarch64/applications/console.c
@@ -31,30 +31,3 @@ static int console_init()
     return status;
 }
 INIT_ENV_EXPORT(console_init);
-
-static int console(int argc, char **argv)
-{
-    rt_err_t result = RT_EOK;
-
-    if (argc > 1)
-    {
-        if (!rt_strcmp(argv[1], "set"))
-        {
-            rt_kprintf("console change to %s\n", argv[2]);
-            rt_console_set_device(argv[2]);
-        }
-        else
-        {
-            rt_kprintf("Unknown command. Please enter 'console' for help\n");
-            result = -RT_ERROR;
-        }
-    }
-    else
-    {
-        rt_kprintf("Usage: \n");
-        rt_kprintf("console set <name>   - change console by name\n");
-        result = -RT_ERROR;
-    }
-    return result;
-}
-MSH_CMD_EXPORT(console, set console name);
